requirements.txt
python-chess==0.22.1
selenium==3.8.1
stockfishpy==1.0.4
